refactor(pdf): estrai pdfKitTheme condiviso (pagina A4, margini, palette, buffer)
Sibling-pattern check pre-push ha segnalato duplicazione reale tra
selfCertificationFormsPlugin.ts (nuovo) e pdfWhitepapersPlugin.ts
(preesistente): stessi valori letterali pagina A4/margini/palette hex
e stesso pattern doc.on('data'|'end'|'error') per bufferizzare pdfkit.
Estratto in build-plugins/shared/pdfKitTheme.ts (AGENTS.md #6), entrambi
i plugin ora importano da lì. Nessun cambio di comportamento (GitNexus
impact su generatePdf: risk LOW, firma invariata, unico caller closeBundle).
